Upgrade to scala 3

+++ app/config/FrontendAppConfig.scala
@@ -48,7 +48,7 @@ class FrontendAppConfig @Inject() (configuration: Configuration, servicesConfig:
   val timeout: Int   = configuration.get[Int]("timeout-dialog.timeout")
   val countdown: Int = configuration.get[Int]("timeout-dialog.countdown")
 
-  val cacheTtl: Int = configuration.get[Int]("mongodb.timeToLiveInSeconds")
+  val cacheTtl: Long = configuration.get[Int]("mongodb.timeToLiveInSeconds")
 
   lazy val dstNewReturnsFrontendEnableFlag: Boolean =
     configuration.getOptional[Boolean]("feature.dstNewReturnsFrontendEnable").getOrElse(false)
